+++ packages/blocks-core/js/libs/wordpress/social-link/test/functionality.blocks-2.e2e.cy.js
@@ -3,6 +3,7 @@
  */
 import {
 	savePage,
+	selectBlockByListView,
 	createPost,
 	appendBlocks,
 	setInnerBlock,
@@ -24,7 +25,7 @@ describe('Social Link Block', () => {
 		appendBlocks(testContent);
 
 		// Select target block
-		cy.getBlock('core/social-link').first().click();
+		selectBlockByListView('WordPress');
 
 		// Block supported is active
 		cy.get('.blockera-extension-block-card').should('be.visible');