+++ packages/blocks-core/js/libs/wordpress/social-links/test/functionality.blocks-2.e2e.cy.js
@@ -4,7 +4,7 @@
 import {
 	savePage,
 	createPost,
-	selectBlock,
+	selectBlockByListView,
 	appendBlocks,
 	setInnerBlock,
 	setParentBlock,
@@ -33,7 +33,7 @@ describe('Social Links Block', () => {
 		openBlockNavigator();
 
 		// Select target block
-		selectBlock('Social Icons');
+		selectBlockByListView('Social Icons');
 
 		cy.getByAriaControls('styles-view').click();