+++ packages/editor/js/extensions/components/test/get-compatible-attributes.spec.js
@@ -117,8 +117,8 @@ describe('syncGroupLayoutFromWp', () => {
 				blockeraFlexLayout: {
 					value: {
 						direction: 'row',
-						justifyContent: '',
-						alignItems: '',
+						flexJustify: '',
+						flexAlign: '',
 					},
 				},
 			},